+++ lib/vagrant/environment.rb
@@ -37,7 +37,6 @@ module Vagrant
         version = VirtualBox.version
         raise Errors::VirtualBoxNotDetected if version.nil?
         raise Errors::VirtualBoxInvalidVersion, :version => version.to_s if version.to_f < 4.0
-        raise Errors::VirtualBoxInvalidOSE, :version => version.to_s if version.to_s.downcase.include?("ose")
       rescue Errors::VirtualBoxNotDetected
         # On 64-bit Windows, show a special error. This error is a subclass
         # of VirtualBoxNotDetected, so libraries which use Vagrant can just